server/app/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
gunicorn==21.2.0
python-dotenv==1.0.1
pyodbc==5.2.0
azure-storage-blob==12.24.0
azure-core==1.32.0
azureml-core==1.58.0.post1
pyspark==3.5.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
starlette==0.41.3
python-multipart==0.0.9